refactor api context
